Remove double-press on dead keys to write one spaced variant
This behavior was an idea I had to make it ‘easier’ to type the spaced
variants, but it just made things more difficult and cumbersome (such
as using markdown codeblocks properly).
If you want to use the spaced variants, use the standard way – press the
dead key followed by the spacebar.